+++ cfdm/test/test_read_write.py
@@ -208,7 +208,7 @@ class read_writeTest(unittest.TestCase):
             g = cfdm.read(tmpfile)
             self.assertEqual(len(g), 1)
             g = g[0]
-            self.assertTrue(f.equals(g, verbose=3))
+            self.assertTrue(f.equals(g))
 
     @unittest.skipIf(True, "Flakey")
     def test_write_netcdf_mode(self):
@@ -533,7 +533,7 @@ class read_writeTest(unittest.TestCase):
         for fmt in self.netcdf_fmts:
             cfdm.write(f, tmpfile, fmt=fmt)
             g = cfdm.read(tmpfile)[0]
-            self.assertTrue(f.equals(g, verbose=3))
+            self.assertTrue(f.equals(g))
 
     def test_read_mask(self):
         """Test reading and writing of netCDF with masked data."""
@@ -695,17 +695,13 @@ class read_writeTest(unittest.TestCase):
         with self.assertRaises(ReadError):
             cfdm.read(tmpfilec2)[0]
 
-        self.assertTrue(f0.equals(f, verbose=3))
+        self.assertTrue(f0.equals(f))
 
         self.assertTrue(
-            f.construct("grid_latitude").equals(
-                c.construct("grid_latitude"), verbose=3
-            )
+            f.construct("grid_latitude").equals(c.construct("grid_latitude"))
         )
         self.assertTrue(
-            f0.construct("grid_latitude").equals(
-                c.construct("grid_latitude"), verbose=3
-            )
+            f0.construct("grid_latitude").equals(c.construct("grid_latitude"))
         )
 
         with self.assertRaises(DatasetTypeError):
@@ -768,8 +764,8 @@ class read_writeTest(unittest.TestCase):
 
         for i in range(0, n):
             j = i + n
-            self.assertTrue(fN[i].data.equals(fN[j].data, verbose=3))
-            self.assertTrue(fN[j].data.equals(fN[i].data, verbose=3))
+            self.assertTrue(fN[i].data.equals(fN[j].data))
+            self.assertTrue(fN[j].data.equals(fN[i].data))
 
         # Check that netCDF4 and h5netcdf give the same results
         for i, j in zip(fN, fH):
@@ -791,7 +787,7 @@ class read_writeTest(unittest.TestCase):
                         for i, j in zip(
                             cfdm.read(tmpfile1), cfdm.read(tmpfile0)
                         ):
-                            self.assertTrue(i.equals(j, verbose=3))
+                            self.assertTrue(i.equals(j))
 
     def test_read_write_Conventions(self):
         """Test the `Conventions` keyword argument to `write`."""
@@ -849,9 +845,9 @@ class read_writeTest(unittest.TestCase):
         for n, f in enumerate(a):
             f.set_property("test_id", str(n))
 
-        cfdm.write(a, tmpfile, verbose=1)
+        cfdm.write(a, tmpfile)
 
-        f = cfdm.read(tmpfile, verbose=1)
+        f = cfdm.read(tmpfile)
 
         self.assertEqual(len(a), len(f))
 
@@ -873,13 +869,13 @@ class read_writeTest(unittest.TestCase):
         e = cfdm.read(tmpfile)
         self.assertTrue(len(e), 10)
 
-        e = cfdm.read(tmpfile, domain=True, verbose=1)
+        e = cfdm.read(tmpfile, domain=True)
         self.assertEqual(len(e), 1)
         e = e[0]
         self.assertIsInstance(e, cfdm.Domain)
-        self.assertTrue(e.equals(e.copy(), verbose=3))
-        self.assertTrue(d.equals(e, verbose=3))
-        self.assertTrue(e.equals(d, verbose=3))
+        self.assertTrue(e.equals(e.copy()))
+        self.assertTrue(d.equals(e))
+        self.assertTrue(e.equals(d))
 
         # 1 field and 1 domain
         cfdm.write([f, d], tmpfile)
@@ -887,9 +883,9 @@ class read_writeTest(unittest.TestCase):
         self.assertTrue(len(g), 1)
         g = g[0]
         self.assertIsInstance(g, cfdm.Field)
-        self.assertTrue(g.equals(f, verbose=3))
+        self.assertTrue(g.equals(f))
 
-        e = cfdm.read(tmpfile, domain=True, verbose=1)
+        e = cfdm.read(tmpfile, domain=True)
         self.assertEqual(len(e), 1)
         e = e[0]
         self.assertIsInstance(e, cfdm.Domain)
@@ -900,9 +896,9 @@ class read_writeTest(unittest.TestCase):
         self.assertTrue(len(g), 1)
         g = g[0]
         self.assertIsInstance(g, cfdm.Field)
-        self.assertTrue(g.equals(f, verbose=3))
+        self.assertTrue(g.equals(f))
 
-        e = cfdm.read(tmpfile, domain=True, verbose=1)
+        e = cfdm.read(tmpfile, domain=True)
         self.assertEqual(len(e), 2)
         self.assertIsInstance(e[0], cfdm.Domain)
         self.assertIsInstance(e[1], cfdm.Domain)
@@ -916,7 +912,7 @@ class read_writeTest(unittest.TestCase):
         g = cfdm.read(tmpfile)
 
         self.assertEqual(len(g), 1)
-        self.assertTrue(g[0].equals(f, verbose=3))
+        self.assertTrue(g[0].equals(f))
 
     def test_write_scalar_domain_ancillary(self):
         """Test the writing of a file with a scalar domain ancillary."""
@@ -1656,6 +1652,63 @@ class read_writeTest(unittest.TestCase):
         x = g.dimension_coordinate("longitude")
         self.assertEqual(x.data.nc_dataset_chunksizes(), (4,))
         self.assertEqual(x.bounds.data.nc_dataset_chunksizes(), (2, 2))
+
+    def test_write_hdf5_consolidated_metadata(self):
+        """Test cfdm.write hdf5_consolidated_metadata keyword."""
+        import pyfive
+
+        f = self.f0.copy()
+        f.nc_set_variable("/forecast/model/q")
+
+        # Non-consolidated
+        cfdm.write(f, tmpfile, hdf5_consolidated_metadata=False)
+        self.assertFalse(pyfive.File(tmpfile).consolidated_metadata)
+
+        cfdm.write(
+            f,
+            tmpfile,
+            hdf5_consolidated_metadata=True,
+            h5py_options={"meta_block_size": 4096},  # Small!
+        )
+        self.assertFalse(pyfive.File(tmpfile).consolidated_metadata)
+
+        cfdm.write(
+            f,
+            tmpfile,
+            hdf5_consolidated_metadata=True,
+            hdf5_expansion_factor=0.5,
+        )
+        self.assertFalse(pyfive.File(tmpfile).consolidated_metadata)
+
+        # Consolidated
+        cfdm.write(f, tmpfile)
+        self.assertTrue(pyfive.File(tmpfile).consolidated_metadata)
+
+        cfdm.write(
+            f,
+            tmpfile,
+            hdf5_consolidated_metadata=True,
+            h5py_options={"meta_block_size": 2**20},  # Large!
+        )
+        self.assertTrue(pyfive.File(tmpfile).consolidated_metadata)
+
+        # Consolidated, lots of variables
+        cfdm.write([f] * 100, tmpfile)
+        self.assertTrue(pyfive.File(tmpfile).consolidated_metadata)
+
+    def test_write_hdf5_expansion_factor(self):
+        """Test cfdm.write hdf5_expansion_factor keyword."""
+        import pyfive
+
+        f = self.f0.copy()
+
+        # Consolidated (large expansion factor)
+        cfdm.write(f, tmpfile, hdf5_expansion_factor=10)
+        self.assertTrue(pyfive.File(tmpfile).consolidated_metadata)
+
+        # Non-consolidated (small expansion factor)
+        cfdm.write(f, tmpfile, hdf5_expansion_factor=0.001)
+        self.assertFalse(pyfive.File(tmpfile).consolidated_metadata)
 
 
 if __name__ == "__main__":